Moved and renamed methods
